requirements.txt
imagehash
pillow
numpy
argparse
psycopg2
python-dotenv